Cleanup

+++ rbtree.rs
@@ -617,140 +617,140 @@ mod tests {
     use item_store::{CopyStore, ShareStore};
 
     #[test]
-    fn test_insert_copy() { Test::test_insert(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new()); }
+    fn test_insert_copy() { Test::test_insert(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new()); }
 
     #[test]
-    fn test_insert_overwrite_copy() { Test::test_insert_overwrite(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new()); }
+    fn test_insert_overwrite_copy() { Test::test_insert_overwrite(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new()); }
 
     #[test]
-    fn test_remove_copy() { Test::test_remove(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new()); }
+    fn test_remove_copy() { Test::test_remove(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new()); }
 
     #[bench]
     fn bench_insert_copy_10(bh: &mut BenchHarness) {
-        Test::bench_insert(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 10, bh);
+        Test::bench_insert(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 10, bh);
     }
 
     #[bench]
     fn bench_insert_copy_100(bh: &mut BenchHarness) {
-        Test::bench_insert(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 100, bh);
+        Test::bench_insert(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 100, bh);
     }
 
     #[bench]
     fn bench_insert_copy_1000(bh: &mut BenchHarness) {
-        Test::bench_insert(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 1000, bh);
+        Test::bench_insert(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 1000, bh);
     }
 
     #[bench]
     fn bench_insert_copy_50000(bh: &mut BenchHarness) {
-        Test::bench_insert(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 50000, bh);
+        Test::bench_insert(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 50000, bh);
     }
 
     #[bench]
     fn bench_find_copy_10(bh: &mut BenchHarness) {
-        Test::bench_find(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 10, bh);
+        Test::bench_find(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 10, bh);
     }
 
     #[bench]
     fn bench_find_copy_100(bh: &mut BenchHarness) {
-        Test::bench_find(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 100, bh);
+        Test::bench_find(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 100, bh);
     }
 
     #[bench]
     fn bench_find_copy_1000(bh: &mut BenchHarness) {
-        Test::bench_find(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 1000, bh);
+        Test::bench_find(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 1000, bh);
     }
 
     #[bench]
     fn bench_find_copy_50000(bh: &mut BenchHarness) {
-        Test::bench_find(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 50000, bh);
+        Test::bench_find(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 50000, bh);
     }
 
     #[bench]
     fn bench_remove_copy_10(bh: &mut BenchHarness) {
-        Test::bench_remove(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 10, bh);
+        Test::bench_remove(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 10, bh);
     }
 
     #[bench]
     fn bench_remove_copy_100(bh: &mut BenchHarness) {
-        Test::bench_remove(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 100, bh);
+        Test::bench_remove(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 100, bh);
     }
 
     #[bench]
     fn bench_remove_copy_1000(bh: &mut BenchHarness) {
-        Test::bench_remove(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 1000, bh);
+        Test::bench_remove(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 1000, bh);
     }
 
     #[bench]
     fn bench_remove_copy_50000(bh: &mut BenchHarness) {
-        Test::bench_remove(RedBlackTree::<uint, uint, CopyStore<uint, uint>>::new(), 50000, bh);
+        Test::bench_remove(RedBlackTree::<u64, u64, CopyStore<u64, u64>>::new(), 50000, bh);
     }
 
     #[test]
-    fn test_insert_shared() { Test::test_insert(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new()); }
+    fn test_insert_shared() { Test::test_insert(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new()); }
 
     #[test]
-    fn test_insert_overwrite_shared() { Test::test_insert_overwrite(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new()); }
+    fn test_insert_overwrite_shared() { Test::test_insert_overwrite(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new()); }
 
     #[test]
-    fn test_remove_shared() { Test::test_remove(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new()); }
+    fn test_remove_shared() { Test::test_remove(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new()); }
 
     #[bench]
     fn bench_insert_shared_10(bh: &mut BenchHarness) {
-        Test::bench_insert(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 10, bh);
+        Test::bench_insert(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 10, bh);
     }
 
     #[bench]
     fn bench_insert_shared_100(bh: &mut BenchHarness) {
-        Test::bench_insert(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 100, bh);
+        Test::bench_insert(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 100, bh);
     }
 
     #[bench]
     fn bench_insert_shared_1000(bh: &mut BenchHarness) {
-        Test::bench_insert(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 1000, bh);
+        Test::bench_insert(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 1000, bh);
     }
 
     #[bench]
     fn bench_insert_shared_50000(bh: &mut BenchHarness) {
-        Test::bench_insert(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 50000, bh);
+        Test::bench_insert(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 50000, bh);
     }
 
     #[bench]
     fn bench_find_shared_10(bh: &mut BenchHarness) {
-        Test::bench_find(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 10, bh);
+        Test::bench_find(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 10, bh);
     }
 
     #[bench]
     fn bench_find_shared_100(bh: &mut BenchHarness) {
-        Test::bench_find(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 100, bh);
+        Test::bench_find(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 100, bh);
     }
 
     #[bench]
     fn bench_find_shared_1000(bh: &mut BenchHarness) {
-        Test::bench_find(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 1000, bh);
+        Test::bench_find(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 1000, bh);
     }
 
     #[bench]
     fn bench_find_shared_50000(bh: &mut BenchHarness) {
-        Test::bench_find(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 50000, bh);
+        Test::bench_find(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 50000, bh);
     }
 
     #[bench]
     fn bench_remove_shared_10(bh: &mut BenchHarness) {
-        Test::bench_remove(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 10, bh);
+        Test::bench_remove(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 10, bh);
     }
 
     #[bench]
     fn bench_remove_shared_100(bh: &mut BenchHarness) {
-        Test::bench_remove(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 100, bh);
+        Test::bench_remove(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 100, bh);
     }
 
     #[bench]
     fn bench_remove_shared_1000(bh: &mut BenchHarness) {
-        Test::bench_remove(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 1000, bh);
+        Test::bench_remove(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 1000, bh);
     }
 
     #[bench]
     fn bench_remove_shared_50000(bh: &mut BenchHarness) {
-        Test::bench_remove(RedBlackTree::<uint, uint, ShareStore<uint, uint>>::new(), 50000, bh);
+        Test::bench_remove(RedBlackTree::<u64, u64, ShareStore<u64, u64>>::new(), 50000, bh);
     }
 }
